Allow CommentStore to handle comment truncation, when available
Depends-On: I92954c922514271d774518d6a6c28a01f33c88c2

+++ blockDisabledAccounts.php
@@ -79,7 +79,12 @@ class BlockDisabledAccounts extends Maintenance {
 
 		$block->setTarget( $user );
 		$block->setBlocker( $scriptUser );
-		$block->mReason = $wgContLang->truncate( $reason, 255 );
+		if ( class_exists( CommentStore::class ) ) {
+			// CommentStore handles truncation
+			$block->mReason = $reason;
+		} else {
+			$block->mReason = $wgContLang->truncate( $reason, 255 );
+		}
 		$block->mExpiry = 'infinity';
 		$block->prevents( 'sendemail', true );
 		$block->prevents( 'editownusertalk', true );
